review fixes

+++ app/src/main/java/ru/netology/nmedia/repository/PostRepositoryImpl.kt
@@ -6,6 +6,7 @@ import ru.netology.nmedia.api.PostsApi
 import ru.netology.nmedia.dao.PostDao
 import ru.netology.nmedia.dto.Post
 import ru.netology.nmedia.entity.PostEntity
+import ru.netology.nmedia.service.ApiError
 
 
 class PostRepositoryImpl
@@ -16,26 +17,72 @@ class PostRepositoryImpl
     }
 
     override suspend fun getAllAsyncRetrofit() {
-        val posts = PostsApi.service.getAll()
-        dao.insert(posts.map { PostEntity.fromDto(it) })
+        try {
+            val response = PostsApi.service.getAll()
+            if (!response.isSuccessful) {
+                throw ApiError(response.code(), response.message())
+            }
+            dao.insert(response.body()!!.map { PostEntity.fromDto(it) })
+
+        } catch (ex: Exception) {
+            throw ex
+        }
     }
 
     override suspend fun likeByIdRetrofit(id: Long) {
-        dao.likeById(id)
-        PostsApi.service.likeById(id)
+
+        try {
+            val response = PostsApi.service.likeById(id)
+            if (!response.isSuccessful) {
+                throw ApiError(response.code(), response.message())
+            }
+            dao.likeById(id)
+
+        } catch (ex: Exception) {
+            throw ex
+        }
     }
 
     override suspend fun saveRetrofit(post: Post) {
-        dao.save(PostEntity.fromDto(post))
+
+        try {
+            val response = PostsApi.service.save(post)
+            if (!response.isSuccessful) {
+                throw ApiError(response.code(), response.message())
+            }
+
+            val body = response.body() ?: throw ApiError(response.code(), response.message())
+            dao.insert(PostEntity.fromDto(body))
+
+        } catch (ex: Exception) {
+            throw ex
+        }
     }
 
     override suspend fun removeByIdRetrofit(id: Long) {
-        PostsApi.service.removeById(id)
-        dao.removeById(id)
+        try {
+            val response = PostsApi.service.removeById(id)
+            if (!response.isSuccessful) {
+                throw ApiError(response.code(), response.message())
+            }
+            dao.removeById(id)
+        } catch (ex: Exception) {
+            throw ex
+        }
     }
 
     override suspend fun deleteLikeByIdRetrofit(id: Long) {
         dao.likeById(id)
+        try {
+            val response = PostsApi.service.deleteLikeById(id)
+            if (!response.isSuccessful) {
+                throw ApiError(response.code(), response.message())
+                dao.likeById(id)
+            }
+
+        } catch (ex: Exception) {
+            throw ex
+        }
     }
 
 }
